Split-up time-stepping and spatial discretisation
